Refactoring tests, fixing path initialization
