+++ test/.clang-tidy
@@ -8,14 +8,17 @@ Checks:
   - 'performance-*'
   - '-bugprone-easily-swappable-parameters'
   - '-bugprone-unchecked-optional-access'
-  - '-cppcoreguidelines-avoid-magic-numbers'
+  - '-cppcoreguidelines-avoid-c-arrays'
   - '-cppcoreguidelines-avoid-do-while'
+  - '-cppcoreguidelines-avoid-magic-numbers'
   - '-cppcoreguidelines-avoid-non-const-global-variables'
+  - '-cppcoreguidelines-pro-bounds-array-to-pointer-decay'
   - '-cppcoreguidelines-pro-type-reinterpret-cast'
   - '-cppcoreguidelines-pro-type-union-access'
   - '-misc-include-cleaner'
   - '-misc-non-private-member-variables-in-classes'
   - '-misc-use-anonymous-namespace'
+  - '-modernize-avoid-c-arrays'
   - '-modernize-return-braced-init-list'
   - '-modernize-use-trailing-return-type'
 FormatStyle: 'file'
