+++ test/source/scwx/wxtest.cpp
@@ -1,3 +1,4 @@
+#include <scwx/qt/main/application_paths.hpp>
 #include <scwx/util/logger.hpp>
 
 #include <aws/core/Aws.h>
@@ -11,6 +12,9 @@ int main(int argc, char** argv)
 
    scwx::util::Logger::Initialize();
    spdlog::set_level(spdlog::level::debug);
+
+   scwx::qt::main::ApplicationPaths::Initialize();
+   scwx::qt::main::ApplicationPaths::LogErrors();
 
    Aws::SDKOptions awsSdkOptions;
    Aws::InitAPI(awsSdkOptions);
