+++ test/test.cmake
@@ -26,7 +26,8 @@ set(SRC_PROVIDER_TESTS source/scwx/provider/aws_level2_data_provider.test.cpp
                        source/scwx/provider/warnings_provider.test.cpp)
 set(SRC_QT_CONFIG_TESTS source/scwx/qt/config/county_database.test.cpp
                         source/scwx/qt/config/radar_site.test.cpp)
-set(SRC_QT_MAIN_TESTS source/scwx/qt/main/application_paths.test.cpp)
+set(SRC_QT_MAIN_TESTS source/scwx/qt/main/application_paths.test.cpp
+                      source/scwx/qt/main/program_options.test.cpp)
 set(SRC_QT_MANAGER_TESTS source/scwx/qt/manager/settings_manager.test.cpp
                          source/scwx/qt/manager/update_manager.test.cpp)
 set(SRC_QT_MAP_TESTS source/scwx/qt/map/map_provider.test.cpp)
